+++ src/Backends/REFPROP/REFPROPMixtureBackend.h
@@ -148,12 +148,16 @@ class REFPROPMixtureBackend : public AbstractState
     */
     void update(CoolProp::input_pairs, double value1, double value2);
 
+    void update_Qmass_pair(CoolProp::input_pairs pair, double v1, double v2) override;
+
     /**
      * @brief Update the state, while providing guess values
      */
     void update_with_guesses(CoolProp::input_pairs, double value1, double value2, const GuessesStructure& guesses);
 
     CoolPropDbl calc_molar_mass(void);
+
+    PhaseMolarMasses calc_phase_molar_masses() override;
 
     void check_loaded_fluid(void);